+++ src/dom/v7400/document.rs
@@ -4,7 +4,7 @@ use crate::{
     dom::v7400::{
         connection::ConnectionsCache,
         definition::DefinitionsCache,
-        object::{scene::SceneHandle, ObjectsCache},
+        object::{scene::SceneHandle, ObjectHandle, ObjectsCache},
     },
     tree::v7400::Tree,
 };
@@ -33,18 +33,25 @@ impl Document {
     }
 
     /// Returns a reference to the objects cache.
-    pub(crate) fn objects(&self) -> &ObjectsCache {
+    pub(crate) fn objects_cache(&self) -> &ObjectsCache {
         &self.objects
     }
 
     /// Returns a reference to the connections cache.
-    pub(crate) fn connections(&self) -> &ConnectionsCache {
+    pub(crate) fn connections_cache(&self) -> &ConnectionsCache {
         &self.connections
     }
 
     /// Returns a reference to the object template definitions.
-    pub(crate) fn definitions(&self) -> &DefinitionsCache {
+    pub(crate) fn definitions_cache(&self) -> &DefinitionsCache {
         &self.definitions
+    }
+
+    /// Returns an iterator of all object nodes.
+    pub fn objects(&self) -> impl Iterator<Item = ObjectHandle<'_>> {
+        self.objects
+            .object_node_ids()
+            .map(move |id| id.to_object_handle(self))
     }
 
     /// Returns `Document` object nodes, which have root object ID of scenes.